Read Conquest ‘descriptives’ output files.

Reads Conquest files with descriptive statistics for the estimated latent variables generated by the ‘descriptives’ statement.

Usage

get.dsc(file)

Arguments

file

Character string with the name of the Conquest descriptives file.

Value

A named list of n elements with n being the number of groups for which descriptive statistics were computed. The names of the list are the group names. Each list contains the following elements:

single.values

A data frame containing the group name, dimension names, the number of observations, mean, standard deviation and variance for each of the latent dimensions. If the file contains descriptive statistics for plausible values, the number of rows in the data frame corresponds to the number of plausible values.

aggregates

A data frame containing the group name, dimension names and aggregated statistics for the mean, standard deviation and variance for each of the latent dimensions as well as (in a separate row) their standard errors.

References

See pp. 162 of Wu, M.L., Adams, R.J., Wilson, M.R., & Haldane, S.A. (2007). ACER ConQuest Version 2.0. Generalised Item Response Modeling Software. Camberwell, Victoria: ACER Press.

Examples

file <- system.file("extdata", "twodim_pvl.dsc", package = "eatModel")
dscPV<- get.dsc(file)
#> Found 1 group(s) and 2 dimension(s) in 'C:/Users/grewered/AppData/Local/Temp/Rtmp4CCI6y/temp_libpath3d80217db35/eatModel/extdata/twodim_pvl.dsc'
#> ℹ All Students
file <- system.file("extdata", "twodim_wle.dsc", package = "eatModel")
dscWL<- get.dsc(file)
#> Found 1 group(s) and 2 dimension(s) in 'C:/Users/grewered/AppData/Local/Temp/Rtmp4CCI6y/temp_libpath3d80217db35/eatModel/extdata/twodim_wle.dsc'
#> ℹ All Students